+++ adapters/src/main/java/io/realm/RealmBaseAdapter.java
@@ -16,10 +16,8 @@
 
 package io.realm;
 
-import android.content.Context;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
-import android.view.LayoutInflater;
 import android.widget.BaseAdapter;
 
 /**
@@ -33,25 +31,18 @@ import android.widget.BaseAdapter;
  * closed. Trying to access Realm objects will at this point also result in a {@code IllegalStateException}.
  */
 public abstract class RealmBaseAdapter<T extends RealmModel> extends BaseAdapter {
-
-    protected LayoutInflater inflater;
     @Nullable
     protected OrderedRealmCollection<T> adapterData;
-    @NonNull
-    protected Context context;
-    private final RealmChangeListener<? extends BaseRealm> listener;
+    private final RealmChangeListener<OrderedRealmCollection<T>> listener;
 
-    public RealmBaseAdapter(@NonNull Context context, @Nullable OrderedRealmCollection<T> data) {
-        //noinspection ConstantConditions
-        if (context == null) {
-            throw new IllegalArgumentException("Context cannot be null");
-        }
-        this.context = context;
+    public RealmBaseAdapter(@Nullable OrderedRealmCollection<T> data) {
+        if (data != null && !data.isManaged())
+            throw new IllegalStateException("Only use this adapter with managed list, " +
+                    "for un-managed lists you can just use the BaseAdapter");
         this.adapterData = data;
-        this.inflater = LayoutInflater.from(context);
-        this.listener = new RealmChangeListener<BaseRealm>() {
+        this.listener = new RealmChangeListener<OrderedRealmCollection<T>>() {
             @Override
-            public void onChange(BaseRealm results) {
+            public void onChange(OrderedRealmCollection<T> results) {
                 notifyDataSetChanged();
             }
         };
@@ -63,11 +54,13 @@ public abstract class RealmBaseAdapter<T extends RealmModel> extends BaseAdapter
 
     private void addListener(@NonNull OrderedRealmCollection<T> data) {
         if (data instanceof RealmResults) {
-            RealmResults realmResults = (RealmResults) data;
-            realmResults.realm.handlerController.addChangeListenerAsWeakReference(listener);
+            RealmResults<T> results = (RealmResults<T>) data;
+            //noinspection unchecked
+            results.addChangeListener((RealmChangeListener) listener);
         } else if (data instanceof RealmList) {
-            RealmList realmList = (RealmList) data;
-            realmList.realm.handlerController.addChangeListenerAsWeakReference(listener);
+            RealmList<T> list = (RealmList<T>) data;
+            //noinspection unchecked
+            list.addChangeListener((RealmChangeListener) listener);
         } else {
             throw new IllegalArgumentException("RealmCollection not supported: " + data.getClass());
         }
@@ -75,11 +68,13 @@ public abstract class RealmBaseAdapter<T extends RealmModel> extends BaseAdapter
 
     private void removeListener(@NonNull OrderedRealmCollection<T> data) {
         if (data instanceof RealmResults) {
-            RealmResults realmResults = (RealmResults) data;
-            realmResults.realm.handlerController.removeWeakChangeListener(listener);
+            RealmResults<T> results = (RealmResults<T>) data;
+            //noinspection unchecked
+            results.removeChangeListener((RealmChangeListener) listener);
         } else if (data instanceof RealmList) {
-            RealmList realmList = (RealmList) data;
-            realmList.realm.handlerController.removeWeakChangeListener(listener);
+            RealmList<T> list = (RealmList<T>) data;
+            //noinspection unchecked
+            list.removeChangeListener((RealmChangeListener) listener);
         } else {
             throw new IllegalArgumentException("RealmCollection not supported: " + data.getClass());
         }