+++ tests/src/androidTest/java/io/realm/RealmBaseAdapterTests.java
@@ -70,12 +70,13 @@ public class RealmBaseAdapterTests {
     }
 
     @Test
-    public void testAdapterParameterExceptions() {
+    public void testAdapterUnmanagedParameterExceptions() {
         RealmResults<AllJavaTypes> resultList = realm.where(AllJavaTypes.class).findAll();
+        RealmList<AllJavaTypes> unmanagedRealmList = new RealmList<>(resultList.toArray(new AllJavaTypes[0]));
         try {
-            new ListViewTestAdapter(null, resultList);
-            fail("Should throw exception if context is null");
-        } catch (IllegalArgumentException ignore) {
+            new ListViewTestAdapter(context, unmanagedRealmList);
+            fail("Should throw exception if list is un-managed");
+        } catch (IllegalStateException ignore) {
         }
     }
 
@@ -85,6 +86,7 @@ public class RealmBaseAdapterTests {
         RealmResults<AllJavaTypes> resultList = realm.where(AllJavaTypes.class).findAll();
         resultList.sort(AllJavaTypes.FIELD_STRING);
         ListViewTestAdapter realmAdapter = new ListViewTestAdapter(context, resultList);
+        //noinspection ConstantConditions
         assertEquals(resultList.first().getFieldString(), realmAdapter.getItem(0).getFieldString());
         assertEquals(resultList.size(), realmAdapter.getCount());
 
@@ -92,6 +94,7 @@ public class RealmBaseAdapterTests {
         AllJavaTypes allTypes = realm.createObject(AllJavaTypes.class, TEST_DATA_SIZE);
         allTypes.setFieldString("test data " + TEST_DATA_SIZE);
         realm.commitTransaction();
+        //noinspection ConstantConditions
         assertEquals(resultList.last().getFieldString(), realmAdapter.getItem(realmAdapter.getCount() - 1).getFieldString());
         assertEquals(resultList.size(), realmAdapter.getCount());
 
@@ -135,12 +138,15 @@ public class RealmBaseAdapterTests {
         RealmResults<AllJavaTypes> resultList = realm.where(AllJavaTypes.class).findAll();
         resultList.sort(AllJavaTypes.FIELD_STRING, Sort.DESCENDING);
         ListViewTestAdapter realmAdapter = new ListViewTestAdapter(context, resultList);
+        //noinspection ConstantConditions
         assertEquals(resultList.first().getFieldString(), realmAdapter.getItem(0).getFieldString());
         assertEquals(resultList.size(), realmAdapter.getCount());
 
         resultList.sort(AllJavaTypes.FIELD_STRING);
 
+        //noinspection ConstantConditions
         assertEquals(resultList.last().getFieldString(), realmAdapter.getItem(resultList.size() - 1).getFieldString());
+        //noinspection ConstantConditions
         assertEquals(resultList.get(TEST_DATA_SIZE / 2).getFieldString(), realmAdapter.getItem(TEST_DATA_SIZE / 2).getFieldString());
         assertEquals(resultList.size(), realmAdapter.getCount());
     }
@@ -159,8 +165,10 @@ public class RealmBaseAdapterTests {
         RealmResults<AllJavaTypes> resultList = realm.where(AllJavaTypes.class).findAll();
         ListViewTestAdapter realmAdapter = new ListViewTestAdapter(context, resultList);
 
+        //noinspection ConstantConditions
         assertEquals(resultList.get(0).getFieldString(), realmAdapter.getItem(0).getFieldString());
         assertEquals(resultList.size(), realmAdapter.getCount());
+        //noinspection ConstantConditions
         assertEquals(resultList.last().getFieldString(), realmAdapter.getItem(resultList.size() - 1).getFieldString());
     }