+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+## 2.0.0 (YYYY-MM-DD)
+
+Works with Realm Java 3.x
+
+### Breaking changes
+
+* Removed `RealmBaseAdapter(@Nonnull Context context, @Nullable OrderedRealmCollection<T> data)`.
+* Removed `RealmRecyclerViewAdapter(@NonNull Context context, @Nullable OrderedRealmCollection<T> data, boolean autoUpdate)`.
+* Removed `RealmBaseAdapter.inflater`.
+* Removed `RealmBaseAdapter.context`.
+* Removed `RealmRecyclerViewAdapter.inflater`.
+* Removed `RealmRecyclerViewAdapter.contex`.
+
+### Enhancements
+
+* Added fine grained notification support to `RealmRecyclerViewAdapter`.
+
+## 1.5.0
+
+### Deprecated
+
+* `RealmBaseAdapter(@Nonnull Context context, @Nullable OrderedRealmCollection<T> data)`
+use `RealmBaseAdapter(@Nullable OrderedRealmCollection<T> data)` instead
+
+* `RealmRecyclerViewAdapter(@NonNull Context context, @Nullable OrderedRealmCollection<T> data, boolean autoUpdate)`
+use `RealmRecyclerViewAdapter(@Nullable OrderedRealmCollection<T> data, boolean autoUpdate)` instead.
+
+### Credits
+
+* Thanks to Bhargav Mogra (@bhargavms) for identifying the unnecessary dependency on context and cleaning up.
+
+
 ## 1.4.1
 
 ### Bug fixes